fixed all routs to properly authonticate the owner

+++ routes/people.js
@@ -11,16 +11,29 @@ const router = express.Router()
 
 // list all people
 router.get('/', authenticate, validate, async (req, res) => {
-  const collection = await Person.find()
+  const collection = await Person.find({owner: req.user._id}).select('-gifts')
   res.send({ data: collection })
 })
 
 // get details for a person
 router.get('/:id', authenticate, validate, async (req, res, next) => {
   try {
-    const document = await Person.findById(req.params.id).populate("gifts")
-    if (!document) throw new ResourceNotFoundError(`We could not find a car with id: ${req.params.id}`)
-    res.json({ data:document })
+    // const person = await Person.findById(req.params.id)
+    const document = await Person.findById(req.params.id)//.populate("gifts")
+    if (JSON.stringify(document.owner) === JSON.stringify(req.user._id)) {
+      if (!document) throw new ResourceNotFoundError(`We could not find a car with id: ${req.params.id}`)
+      res.json({ data:document })
+    } else {
+        res.status(500).send({
+          errors: [
+            {
+              status: '400',
+              title: 'Not authorized',
+              description: 'User is not authorized to modify this person',
+            },
+          ],
+        })
+      }
   } catch (err) {
     next()
   }
@@ -28,6 +41,7 @@ router.get('/:id', authenticate, validate, async (req, res, next) => {
 
 // create a person
 router.post('/', sanitizeBody, authenticate, validate, async (req, res) => {
+  req.sanitizedBody.owner = req.user._id
   let newDocument = new Person(req.sanitizedBody)
   try {
     await newDocument.save()
@@ -39,7 +53,7 @@ router.post('/', sanitizeBody, authenticate, validate, async (req, res) => {
         {
           status: '500',
           title: 'Server error',
-          description: 'Problem saving document to the database.',
+          description: 'Problem saving document to the database. check the person data format.',
         },
       ],
     })
@@ -50,7 +64,9 @@ const update =
   (overwrite = false) =>
   async (req, res, next) => {
     try {
-      const document = await Person.findByIdAndUpdate(
+      const person = await Person.findById(req.params.id)
+      if (JSON.stringify(person.owner) === JSON.stringify(req.user._id)) {
+        const document = await Person.findByIdAndUpdate(
         req.params.id,
         req.sanitizedBody,
         {
@@ -59,8 +75,19 @@ const update =
           runValidators: true,
         }
       )
-      if (!document) throw new ResourceNotFoundError(`We could not find a car with id: ${req.params.id}`)
+      if (!document) throw new ResourceNotFoundError(`We could not find a person with id: ${req.params.id}`)
       res.send({ data: document })
+      } else {
+        res.status(500).send({
+          errors: [
+            {
+              status: '400',
+              title: 'Not authorized',
+              description: 'User is not authorized to modify this person',
+            },
+          ],
+        })
+      }
     } catch (err) {
       next()
     }
@@ -71,9 +98,22 @@ router.patch('/:id', sanitizeBody, authenticate, validate, update(false)) // upd
 // remove a person
 router.delete('/:id', authenticate, validate, async (req, res, next) => {
   try {
-    const document = await Person.findByIdAndRemove(req.params.id)
-    if (!document) throw new ResourceNotFoundError(`We could not find a person with id: ${req.params.id}`)
-    res.send({ data:document })
+    const person = await Person.findById(req.params.id)
+    if (JSON.stringify(person.owner) === JSON.stringify(req.user._id)) {
+      const document = await Person.findByIdAndRemove(req.params.id)
+      if (!document) throw new ResourceNotFoundError(`We could not find a person with id: ${req.params.id}`)
+      res.send({ data:document})
+    } else {
+        res.status(500).send({
+          errors: [
+            {
+              status: '400',
+              title: 'Not authorized',
+              description: 'User is not authorized to modify this person',
+            },
+          ],
+        })
+      }
   } catch (err) {
     next()
   }
